# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Costs Service",
    "description": "Cost tracking and budget management service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3011"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "CostName": {
        "type": "string",
        "description": "Provider cost identifier ({provider}-{service-or-model}-{unit-type})",
        "example": "anthropic-sonnet-4.5-tokens-input"
      },
...
      "OrgId": {
        "type": "string",
        "format": "uuid",
        "description": "Internal org UUID from client-service",
        "example": "550e8400-e29b-41d4-a716-446655440000"
      },
      "UserId": {
        "type": "string",
        "format": "uuid",
        "description": "Internal user UUID from client-service",
        "example": "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
      },
      "RunId": {
        "type": "string",
        "format": "uuid",
        "description": "Run UUID from runs-service identifying the current execution",
        "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      },
      "CampaignId": {
        "type": "string",
        "description": "Campaign identifier injected by workflow-service (optional)",
        "example": "camp_abc123"
      },
      "BrandId": {
        "type": "string",
        "description": "Brand identifier(s) injected by workflow-service (optional). Supports comma-separated UUIDs for multi-brand campaigns, e.g. 'uuid1,uuid2,uuid3'.",
        "example": "550e8400-e29b-41d4-a716-446655440000,6ba7b810-9dad-11d1-80b4-00c04fd430c8"
      },
      "WorkflowSlug": {
        "type": "string",
        "description": "Workflow slug injected by workflow-service (optional)",
        "example": "lead-enrichment-v2"
      },
      "FeatureSlug": {
        "type": "string",
        "description": "Feature slug for tracking which feature triggered the request (optional)",
        "example": "press-outreach"
      },
      "AudienceId": {
        "type": "string",
        "description": "Priority audience identifier chosen by campaign-service for cost attribution, injected by workflow-service (optional)",
        "example": "aud_abc123"
      },
...
      "ProviderCost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "providerDomain": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "planTier": {
            "type": "string"
          },
          "billingCycle": {
            "type": "string"
          },
          "costPerUnitInUsdCents": {
            "type": "string"
          },
          "pricingRegime": {
            "type": "string",
            "nullable": true,
            "description": "Time-of-day pricing regime this price belongs to ('peak' or 'off-peak'), or null when the provider charges one rate at every hour. The cost name carries the same segment.",
            "example": "peak"
          },
          "regimeHoursUtc": {
            "type": "string",
            "nullable": true,
            "description": "UTC windows during which this regime is in force, as comma-separated half-open HH:MM-HH:MM ranges. Null when pricingRegime is null. A provider's regimes partition the day, so exactly one cost name matches any instant.",
            "example": "01:00-04:00,06:00-10:00"
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "provider",
          "providerDomain",
          "type",
          "unit",
          "planTier",
          "billingCycle",
          "costPerUnitInUsdCents",
          "pricingRegime",
          "regimeHoursUtc",
          "effectiveFrom",
          "createdAt",
          "updatedAt"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "properties": {
              "formErrors": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "fieldErrors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "required": [
              "formErrors",
              "fieldErrors"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "PutProviderCostBody": {
        "type": "object",
        "properties": {
          "costPerUnitInUsdCents": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "provider": {
            "type": "string"
          },
          "providerDomain": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "planTier": {
            "type": "string"
          },
          "billingCycle": {
            "type": "string"
          },
          "pricingRegime": {
            "type": "string",
            "nullable": true
          },
          "regimeHoursUtc": {
            "type": "string",
            "nullable": true
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "costPerUnitInUsdCents",
          "provider",
          "type",
          "unit",
          "planTier",
          "billingCycle"
        ]
      },
      "DeleteProviderCostResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "number"
          }
        },
        "required": [
          "deleted"
        ]
      },
...
    },
    "parameters": {
      "CostName": {
        "schema": {
          "$ref": "#/components/schemas/CostName"
        },
        "required": true,
        "description": "Provider cost identifier ({provider}-{service-or-model}-{unit-type})",
        "name": "name",
        "in": "path"
      },
...
      "OrgId": {
        "schema": {
          "$ref": "#/components/schemas/OrgId"
        },
        "required": true,
        "description": "Internal org UUID from client-service",
        "name": "x-org-id",
        "in": "header"
      },
      "UserId": {
        "schema": {
          "$ref": "#/components/schemas/UserId"
        },
        "required": true,
        "description": "Internal user UUID from client-service",
        "name": "x-user-id",
        "in": "header"
      },
      "RunId": {
        "schema": {
          "$ref": "#/components/schemas/RunId"
        },
        "required": true,
        "description": "Run UUID from runs-service identifying the current execution",
        "name": "x-run-id",
        "in": "header"
      },
      "CampaignId": {
        "schema": {
          "$ref": "#/components/schemas/CampaignId"
        },
        "required": false,
        "description": "Campaign identifier injected by workflow-service (optional)",
        "name": "x-campaign-id",
        "in": "header"
      },
      "BrandId": {
        "schema": {
          "$ref": "#/components/schemas/BrandId"
        },
        "required": false,
        "description": "Brand identifier(s) injected by workflow-service (optional). Supports comma-separated UUIDs for multi-brand campaigns, e.g. 'uuid1,uuid2,uuid3'.",
        "name": "x-brand-id",
        "in": "header"
      },
      "WorkflowSlug": {
        "schema": {
          "$ref": "#/components/schemas/WorkflowSlug"
        },
        "required": false,
        "description": "Workflow slug injected by workflow-service (optional)",
        "name": "x-workflow-slug",
        "in": "header"
      },
      "FeatureSlug": {
        "schema": {
          "$ref": "#/components/schemas/FeatureSlug"
        },
        "required": false,
        "description": "Feature slug for tracking which feature triggered the request (optional)",
        "name": "x-feature-slug",
        "in": "header"
      },
      "AudienceId": {
        "schema": {
          "$ref": "#/components/schemas/AudienceId"
        },
        "required": false,
        "description": "Priority audience identifier chosen by campaign-service for cost attribution, injected by workflow-service (optional)",
        "name": "x-audience-id",
        "in": "header"
      }
    }
  },
  "paths": {
...
    "/v1/providers-costs": {
      "get": {
        "operationId": "listProvidersCosts",
        "summary": "List all provider costs (resolved via platform plan per provider)",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/CampaignId"
          },
          {
            "$ref": "#/components/parameters/BrandId"
          },
          {
            "$ref": "#/components/parameters/WorkflowSlug"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "$ref": "#/components/parameters/AudienceId"
          }
        ],
        "responses": {
          "200": {
            "description": "List of current provider costs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderCost"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/providers-costs/{name}": {
      "get": {
        "operationId": "getProviderCost",
        "summary": "Get current provider cost (resolved via platform plan)",
        "parameters": [
          {
            "$ref": "#/components/parameters/CostName"
          },
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/CampaignId"
          },
          {
            "$ref": "#/components/parameters/BrandId"
          },
          {
            "$ref": "#/components/parameters/WorkflowSlug"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "$ref": "#/components/parameters/AudienceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Current provider cost",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderCost"
                }
              }
            }
          },
          "404": {
            "description": "Provider cost not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error (e.g. no platform plan configured)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "putProviderCost",
        "summary": "Insert a new price point for a provider cost",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CostName"
          },
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/CampaignId"
          },
          {
            "$ref": "#/components/parameters/BrandId"
          },
          {
            "$ref": "#/components/parameters/WorkflowSlug"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "$ref": "#/components/parameters/AudienceId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutProviderCostBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inserted provider cost",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderCost"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown cost name — cost must already exist in the catalog",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate provider cost (name + plan_tier + billing_cycle + effective_from already exists)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteProviderCost",
        "summary": "Delete all entries for a provider cost",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CostName"
          },
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/CampaignId"
          },
          {
            "$ref": "#/components/parameters/BrandId"
          },
          {
            "$ref": "#/components/parameters/WorkflowSlug"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "$ref": "#/components/parameters/AudienceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Number of deleted entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProviderCostResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Provider cost not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/providers-costs/{name}/history": {
      "get": {
        "operationId": "getProviderCostHistory",
        "summary": "Get all price points for a provider cost",
        "parameters": [
          {
            "$ref": "#/components/parameters/CostName"
          },
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/CampaignId"
          },
          {
            "$ref": "#/components/parameters/BrandId"
          },
          {
            "$ref": "#/components/parameters/WorkflowSlug"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "$ref": "#/components/parameters/AudienceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Price history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderCost"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Provider cost not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/providers-costs/{name}/plans": {
      "get": {
        "operationId": "getProviderCostPlans",
        "summary": "List all known plan options for a provider cost",
        "parameters": [
          {
            "$ref": "#/components/parameters/CostName"
          },
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/CampaignId"
          },
          {
            "$ref": "#/components/parameters/BrandId"
          },
          {
            "$ref": "#/components/parameters/WorkflowSlug"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "$ref": "#/components/parameters/AudienceId"
          }
        ],
        "responses": {
          "200": {
            "description": "All plan/billing cycle combinations for this cost",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderCost"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Provider cost not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}