# src/routes/providers-costs.ts
import { Router } from "express";
import { eq, lte, desc, and } from "drizzle-orm";
import { db } from "../db/index.js";
import { providersCosts, platformCosts } from "../db/schema.js";
import { requireApiKey } from "../middleware/auth.js";
import { PutProviderCostBodySchema } from "../schemas.js";
import { getTraceIdentityHeaders, traceEvent } from "../lib/trace-event.js";

const router = Router();

/** Resolve the current platform cost config for a provider. Returns null if none configured. */
async function getCurrentPlatformCost(provider: string) {
  const now = new Date();
  const [cost] = await db
    .select()
    .from(platformCosts)
    .where(and(eq(platformCosts.provider, provider), lte(platformCosts.effectiveFrom, now)))
    .orderBy(desc(platformCosts.effectiveFrom))
    .limit(1);
  return cost ?? null;
}

// GET /v1/providers-costs — list all current provider costs (resolved via platform plan)
router.get("/v1/providers-costs", async (req, res) => {
  try {
    const now = new Date();

    // 1. Get all current platform costs (latest per provider)
    const allPlatformCosts = await db
      .select()
      .from(platformCosts)
      .where(lte(platformCosts.effectiveFrom, now))
      .orderBy(platformCosts.provider, desc(platformCosts.effectiveFrom));

    const planMap = new Map<string, { planTier: string; billingCycle: string }>();
    for (const pc of allPlatformCosts) {
      if (!planMap.has(pc.provider)) {
        planMap.set(pc.provider, { planTier: pc.planTier, billingCycle: pc.billingCycle });
      }
    }

    // 2. Get all costs where effectiveFrom <= now
    const allCosts = await db
      .select()
      .from(providersCosts)
      .where(lte(providersCosts.effectiveFrom, now))
      .orderBy(providersCosts.name, desc(providersCosts.effectiveFrom));

    // 3. Filter by matching platform cost config, deduplicate per name
    const seen = new Set<string>();
    const current = allCosts.filter((row) => {
      if (seen.has(row.name)) return false;
      const plan = planMap.get(row.provider);
      if (!plan) return false;
      if (row.planTier !== plan.planTier || row.billingCycle !== plan.billingCycle) return false;
      seen.add(row.name);
      return true;
    });

    traceEvent({
      runId: req.headers["x-run-id"] as string | undefined,
      event: "providers_costs.listed",
      detail: `Listed ${current.length} current provider costs resolved against ${planMap.size} active platform provider configs.`,
      identityHeaders: getTraceIdentityHeaders(req),
    });

    res.json(current);
  } catch (err) {
    console.error("[Costs Service] Error listing providers costs:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

// GET /v1/providers-costs/:name — current provider cost for one name (resolved via platform plan)
router.get("/v1/providers-costs/:name", async (req, res) => {
  try {
    const { name } = req.params;
    const now = new Date();

    // 1. Find the provider for this cost name
    const [anyCost] = await db
      .select({ provider: providersCosts.provider })
      .from(providersCosts)
      .where(eq(providersCosts.name, name))
      .limit(1);

    if (!anyCost) {
      res.status(404).json({ error: "Provider cost not found" });
      return;
    }

    // 2. Get current platform cost config for this provider
    const platformCost = await getCurrentPlatformCost(anyCost.provider);
    if (!platformCost) {
      res.status(500).json({ error: `No platform cost configured for provider '${anyCost.provider}'` });
      return;
    }

    // 3. Get the cost matching our plan
    const [result] = await db
      .select()
      .from(providersCosts)
      .where(
        and(
          eq(providersCosts.name, name),
          eq(providersCosts.planTier, platformCost.planTier),
          eq(providersCosts.billingCycle, platformCost.billingCycle),
          lte(providersCosts.effectiveFrom, now),
        ),
      )
      .orderBy(desc(providersCosts.effectiveFrom))
      .limit(1);

    if (!result) {
      res.status(404).json({
        error: `No cost found for '${name}' on plan '${platformCost.planTier}/${platformCost.billingCycle}'`,
      });
      return;
    }

    traceEvent({
      runId: req.headers["x-run-id"] as string | undefined,
      event: "provider_cost.resolved",
      detail: `Resolved provider cost '${name}' for provider '${result.provider}' using active plan '${platformCost.planTier}/${platformCost.billingCycle}' effective ${result.effectiveFrom.toISOString()}.`,
      identityHeaders: getTraceIdentityHeaders(req),
    });

    res.json(result);
  } catch (err) {
    console.error("[Costs Service] Error getting provider cost:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

// GET /v1/providers-costs/:name/history — all price points for a name (all plans)
router.get("/v1/providers-costs/:name/history", async (req, res) => {
  try {
    const { name } = req.params;

    const history = await db
      .select()
      .from(providersCosts)
      .where(eq(providersCosts.name, name))
      .orderBy(desc(providersCosts.effectiveFrom));

    if (history.length === 0) {
      res.status(404).json({ error: "Provider cost not found" });
      return;
    }

    res.json(history);
  } catch (err) {
    console.error("[Costs Service] Error getting provider cost history:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

// GET /v1/providers-costs/:name/plans — all plan options for a cost name
router.get("/v1/providers-costs/:name/plans", async (req, res) => {
  try {
    const { name } = req.params;
    const now = new Date();

    // Get latest effective price per (plan_tier, billing_cycle) combo
    const allCosts = await db
      .select()
      .from(providersCosts)
      .where(and(eq(providersCosts.name, name), lte(providersCosts.effectiveFrom, now)))
      .orderBy(providersCosts.planTier, providersCosts.billingCycle, desc(providersCosts.effectiveFrom));

    if (allCosts.length === 0) {
      res.status(404).json({ error: "Provider cost not found" });
      return;
    }

    // Deduplicate: keep latest per (planTier, billingCycle)
    const seen = new Set<string>();
    const plans = allCosts.filter((row) => {
      const key = `${row.planTier}:${row.billingCycle}`;
      if (seen.has(key)) return false;
      seen.add(key);
      return true;
    });

    res.json(plans);
  } catch (err) {
    console.error("[Costs Service] Error getting provider cost plans:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

// PUT /v1/providers-costs/:name — insert new price point
router.put("/v1/providers-costs/:name", requireApiKey, async (req, res) => {
  try {
    const { name } = req.params;
    const parsed = PutProviderCostBodySchema.safeParse(req.body);

    if (!parsed.success) {
      res.status(400).json({ error: "Invalid request", details: parsed.error.flatten() });
      return;
    }

    // Reject unknown cost names — only allow new price points for costs that already exist in the catalog
    const [existing] = await db
      .select({ name: providersCosts.name })
      .from(providersCosts)
      .where(eq(providersCosts.name, name))
      .limit(1);

    if (!existing) {
      console.error(`[costs-service] PUT /v1/providers-costs/${name} rejected: cost name not found in catalog`);
      res.status(404).json({
        error: `Unknown cost name '${name}'. Cost must already exist in the catalog. Check GET /v1/providers-costs for available names.`,
      });
      return;
    }

    const { costPerUnitInUsdCents, provider, providerDomain, type, unit, planTier, billingCycle, pricingRegime, regimeHoursUtc, effectiveFrom } = parsed.data;

    const [inserted] = await db
      .insert(providersCosts)
      .values({
        name,
        provider,
        providerDomain: providerDomain ?? null,
        type,
        unit,
        planTier,
        billingCycle,
        pricingRegime: pricingRegime ?? null,
        regimeHoursUtc: regimeHoursUtc ?? null,
        costPerUnitInUsdCents: String(costPerUnitInUsdCents),
        effectiveFrom: effectiveFrom ? new Date(effectiveFrom) : new Date(),
      })
      .returning();

    traceEvent({
      runId: req.headers["x-run-id"] as string | undefined,
      event: "provider_cost.updated",
      detail: `Inserted provider cost '${name}' for provider '${provider}' plan '${planTier}/${billingCycle}' effective ${inserted.effectiveFrom.toISOString()}.`,
      identityHeaders: getTraceIdentityHeaders(req),
    });

    res.json(inserted);
  } catch (err: any) {
    if (err?.code === "23505") {
      res.status(409).json({ error: "Provider cost with this name, plan, billing cycle, and effective_from already exists" });
      return;
    }
    console.error("[Costs Service] Error upserting provider cost:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

// DELETE /v1/providers-costs/:name — delete all entries for a name
router.delete("/v1/providers-costs/:name", requireApiKey, async (req, res) => {
  try {
    const { name } = req.params;

    const deleted = await db
      .delete(providersCosts)
      .where(eq(providersCosts.name, name))
      .returning();

    if (deleted.length === 0) {
      res.status(404).json({ error: "Provider cost not found" });
      return;
    }

    res.json({ deleted: deleted.length });
  } catch (err) {
    console.error("[Costs Service] Error deleting provider cost:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;